requirements.txt
-e .
fastapi>=0.111.0
uvicorn[standard]>=0.30.1
pydantic>=2.7.0
python-dotenv>=1.0.1
sentence-transformers>=3.0.1
faiss-cpu>=1.8.0
whoosh>=2.7.4
duckdb>=0.10.2
pandas>=2.2.2
numpy>=1.26.4
Pillow>=10.3.0
transformers>=4.41.0
torch>=2.2.0
ragas>=0.1.17
matplotlib>=3.8.0
streamlit>=1.35.0
tqdm>=4.66.4
scikit-learn>=1.4.2
rich>=13.7.1
requests>=2.31.0
beautifulsoup4>=4.12.2
pdfminer.six>=20231228
datasets>=2.18.0
openai>=1.30.0
pytest>=8.2.0
pytest-asyncio>=0.23.7
pytest-cov>=5.0.0
mypy>=1.10.0
ruff>=0.4.8
pre-commit>=3.7.0
